Change the way to define the main method for Java
To define the main method of Java from Kotlin if we define the Kotlin class as object is not needed to wrap our main function inside the companion object. In my opinion, this approach is easier to understand if we want to define Java's `main` static method.
But if the the intention of the section is to show how to define any (not the main one) method as static for Java, maybe the `main` method is not the best example.
Also in this case the method parameters ar not well defined as `doWork()` has no parameters.